+++ src/base/vanilla/localQuestTime.js
@@ -1,8 +1,6 @@
 onPage('Quests', function localTime() {
-  eventManager.on('jQuery', () => {
-    const c = $('p:contains("(Paris time zone).")');
-    const raw = $(c[0]).text().replace(/[^0-9\:]/gi, '').split(':').map(Number);
-    const time = luxon.DateTime.fromObject({ hour: raw[0], minute: raw[1], zone: 'Europe/Paris' }).toLocal();
-    c.append(` (${time.toLocaleString(luxon.DateTime.TIME_SIMPLE)} local)`)
+  eventManager.on(':load', () => {
+    const time = luxon.DateTime.fromObject({ hour: 6, minute: 0, zone: 'Europe/Paris' }).toLocal();
+    $('[data-i18n="[html]quests-reset"],[data-i18n="[html]quests-day"]').append(` (${time.toLocaleString(luxon.DateTime.TIME_SIMPLE)} local)`)
   });
 });