+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Version 0.27.1
 1. Fixed card skin shop names looking small
+1. Fixed local reset time
 
 ## Version 0.27.0
 ### New Features